timedcapture/requirements.txt
boto3==1.17.54
botocore==1.20.54
certifi==2018.10.15
configparser==5.0.2
docutils==0.15.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
future==0.18.2
jetson-stats==3.1.0
jmespath==0.10.0
minio==5.0.9
mysql-connector-python==8.0.21
mysql-replication==0.21
numpy==1.19.1
opencv-python==4.5.1.48
protobuf==3.15.8
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
s3transfer==0.4.0
six==1.15.0
urllib3==1.26.4
